+++ lib/src/user_location_layer.dart
@@ -24,11 +24,18 @@ class _MapsPluginLayerState extends State<MapsPluginLayer>
   Marker _locationMarker;
   EventChannel _stream = EventChannel('locationStatusStream');
   var location = Location();
+
+  bool mapLoaded;
+
   StreamSubscription<LocationData> _onLocationChangedStreamSubscription;
+
 
   @override
   void initState() {
     super.initState();
+    setState(() {
+      mapLoaded = false;
+    });
     initialize();
   }
 
@@ -134,6 +141,14 @@ class _MapsPluginLayerState extends State<MapsPluginLayer>
             printLog(
                 "Warning: updateMapLocationOnPositionChange set to true, but no mapController provided: can't move map");
           }
+
+          if (widget.options.zoomToCurrentLocationOnLoad && (!mapLoaded)) {
+            setState(() {
+              mapLoaded = true;
+            });
+            animatedMapMove(
+                _currentLocation, 17, widget.options.mapController, this);
+          }
         });
       });
     }
@@ -190,7 +205,10 @@ class _MapsPluginLayerState extends State<MapsPluginLayer>
                 hoverColor: Colors.blueAccent[200],
                 onTap: () {
                   initialize();
-                  _moveMapToCurrentLocation();
+                  widget.options.mapController.move(
+                    _currentLocation,
+                    17.0,
+                  );
                 },
                 child: widget.options
                             .moveToCurrentLocationFloatingActionButton ==